Refactored child components into their own sub directory for easier import
