+++ components/image/child-components/index.js
@@ -1,0 +1,5 @@
+import { Media, ImageContext } from './media';
+import { Figure } from './figure';
+import { InlineControls } from './inline-controls';
+
+export { Media, ImageContext, Figure, InlineControls };
