+++ components/image/child-components/inline-controls.js
